+++ apps/laboratory/src/pages/library/ethers.tsx
@@ -13,6 +13,7 @@ const modal = createWeb3Modal({
     coinbasePreference: 'smartWalletOnly'
   }),
   caipNetworks: EthersConstants.chains,
+  defaultChain: EthersConstants.chains[1],
   projectId: ConstantsUtil.ProjectId,
   enableAnalytics: true,
   metadata: ConstantsUtil.Metadata,