+++ apps/laboratory/src/pages/library/wagmi.tsx
@@ -7,6 +7,7 @@ import { ThemeStore } from '../../utils/StoreUtil'
 import { getWagmiConfig, WagmiConstantsUtil } from '../../utils/WagmiConstants'
 import { ConstantsUtil } from '../../utils/ConstantsUtil'
 import { WagmiModalInfo } from '../../components/Wagmi/WagmiModalInfo'
+import { mainnet } from 'viem/chains'
 
 const queryClient = new QueryClient()
 
@@ -15,6 +16,7 @@ const wagmiConfig = getWagmiConfig('default')
 const modal = createWeb3Modal({
   wagmiConfig,
   caipNetworks: WagmiConstantsUtil.chains,
+  defaultChain: mainnet,
   projectId: ConstantsUtil.ProjectId,
   enableAnalytics: true,
   metadata: ConstantsUtil.Metadata,