+++ packages/base/adapters/solana/web3js/utils/defaultConfig.ts
@@ -1,6 +1,6 @@
 import '@web3modal/polyfills'
-
-import type { Chain, Metadata, Provider, ProviderType } from '@web3modal/scaffold-utils/solana'
+import type { CaipNetwork } from '@web3modal/common'
+import type { Metadata, Provider, ProviderType } from '@web3modal/scaffold-utils/solana'
 
 declare global {
   interface Window {
@@ -14,9 +14,9 @@ declare global {
   }
 }
 
-export interface ConfigOptions {
+export interface ConfigOptions extends Pick<Provider, 'auth'> {
   projectId?: string
-  chains: Chain[]
+  chains: CaipNetwork[]
   enableInjected?: boolean
   rpcUrl?: string
   defaultChainId?: number
@@ -28,7 +28,7 @@ export function defaultSolanaConfig(options: ConfigOptions) {
 
   let injectedProvider: Provider | undefined = undefined
 
-  const providers: ProviderType = { metadata }
+  const providers: ProviderType = { metadata, auth: options.auth }
 
   function getInjectedProvider() {
     if (injectedProvider) {