+++ packages/base/adapters/solana/web3js/utils/withSolanaNamespace.ts
@@ -1,0 +1,9 @@
+export function withSolanaNamespace<T = string | undefined>(
+  chainId?: T
+): T extends string ? `solana:${string}` : undefined {
+  if (typeof chainId === 'string') {
+    return `solana:${chainId}` as T extends string ? `solana:${string}` : undefined
+  }
+
+  return undefined as T extends string ? `solana:${string}` : undefined
+}