+++ packages/base/adapters/wc/universal-adapter/client.ts
@@ -23,7 +23,7 @@ import {
   type CaipAddress,
   type ChainNamespace
 } from '@web3modal/common'
-import { ProviderUtil } from '../../../utils/ProviderUtil.js'
+import { ProviderUtil } from '../../../utils/store/ProviderUtil.js'
 
 type Metadata = {
   name: string
@@ -59,7 +59,7 @@ export class UniversalAdapterClient {
 
   public chainNamespace: ChainNamespace
 
-  public defaultNetwork: CaipNetwork
+  public defaultNetwork: CaipNetwork | undefined = undefined
 
   public networkControllerClient: NetworkControllerClient