+++ packages/wagmi/exports/index.ts
@@ -1,5 +1,5 @@
 import { AppKit } from '@web3modal/base'
-import type { AppKitOptions, ChainAdapter } from '@web3modal/base'
+import type { AppKitOptions } from '@web3modal/base'
 import { EVMWagmiClient, type AdapterOptions } from '@web3modal/base/adapters/evm/wagmi'
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 import type { Config } from 'wagmi'
@@ -14,11 +14,11 @@ export { authConnector } from '@web3modal/base/adapters/evm/wagmi'
 export { defaultWagmiConfig } from '@web3modal/base/adapters/evm/wagmi'
 
 // -- Setup -------------------------------------------------------------
-type WagmiAppKitOptions = Omit<AppKitOptions, 'adapters' | 'sdkType' | 'sdkVersion'> &
+export type WagmiAppKitOptions = Omit<AppKitOptions, 'adapters' | 'sdkType' | 'sdkVersion'> &
   AdapterOptions<Config>
 
 export function createWeb3Modal(options: WagmiAppKitOptions) {
-  const wagmiAdapter = new EVMWagmiClient() as ChainAdapter
+  const wagmiAdapter = new EVMWagmiClient()
 
   return new AppKit({
     ...options,